requirements.txt
gradio==3.36.1
opencv-python==4.7.0.72
Pillow==10.1.0
accelerate==0.24.1
diffusers==0.23.1
numpy==1.24.4
requests
diffusers==0.23.1
transformers==4.35.2